Clean up macros that allow building with clang
This adds the --with toolchain_clang option which can be passed to
fedpkg, mock, or rpmbuild to allow for building with clang.  It also
changes the conditional parts of the spec that are only used for
building with clang to check the value of the generic %toolchain macro
rather than a macro that is specific to thunderbird.